fix: avatar storage requests on client
